+++ dynamite/weight_solvers.py
@@ -12,8 +12,8 @@ try:
 except ModuleNotFoundError:
     pass
 
+from dynamite import constants
 from dynamite import analysis
-from dynamite import physical_system as physys
 from dynamite import kinematics as dyn_kin
 
 
@@ -47,7 +47,7 @@ class WeightSolver(object):
         if 'CRcut' in self.settings.keys():
             CRcut = self.settings['CRcut']
         self.CRcut = CRcut
-        self.weight_file = f'{self.direc_with_ml}orbit_weights.ecsv'
+        self.weight_file = f'{self.direc_with_ml}{constants.weight_file}'
 
     def solve(self, orblib, ignore_existing_weights=False):
         """Template solve method
@@ -559,8 +559,8 @@ class LegacyWeightSolver(WeightSolver):
         a = self.__read_file_element(fname, [1, 1], [1, 2])
         ngh = np.int64(a[1])  # number of 'observables'
         nobs = np.int64(a[1])
-        nvel = np.int64(a[0])
-        ncon = np.int64(a[0])
+        # nvel = np.int64(a[0])
+        # ncon = np.int64(a[0])
         rows = 3 + np.arange(nobs)  # rows 1- 9
         cols = 3 + np.zeros(nobs, dtype=int)  # skip over text
         fname = self.fname_nn_nnls
@@ -844,9 +844,6 @@ class NNLS(WeightSolver):
         """
         self.logger.info(f"Using WeightSolver: {__class__.__name__}/"
                          f"{self.nnls_solver}")
-        orblib.read_vel_histograms()  # sets orblib.vel_histograms,
-                                      # orblib.intrinsic_masses, and
-                                      # orblib.projected_masses
         if (not ignore_existing_weights) and self.weight_file_exists():
             results = ascii.read(self.weight_file, format='ecsv')
             self.logger.info("NNLS solution read from existing output "
@@ -856,6 +853,9 @@ class NNLS(WeightSolver):
             chi2_kin = results.meta['chi2_kin']
             chi2_kinmap = results.meta['chi2_kinmap']
         else:
+            orblib.read_vel_histograms()  # sets orblib.vel_histograms,
+                                          # orblib.intrinsic_masses, and
+                                          # orblib.projected_masses
             A, b = self.construct_nnls_matrix_and_rhs(orblib)
             if self.nnls_solver=='scipy':
                 try: