+++ docs/more_info/changelog.rst
@@ -5,6 +5,10 @@ Change Log
 ****************
 
 - Improvement: Changed the default BayesLOSVD histogram width and number of velocity bins to be consistent with the data (no oversampling by default).
+- Improvement: Use indicator files for existing orblibs instead of checking multiple files (preparation for coloring and proper motions).
+- Improvement: Replaced staging files by directly checking for orblibs and weight files on disk.
+- Improvement: Significant performance gain when reading existing model weights (eliminated unnecessary reading of orbit library)
+- Bugfix: fix a bug in beta_plot and anistotropy_single, subtract first moments, plot only beta_r, enable log x-axis, plot pc/kpc scale on upper x-axis
 - New feature: DYNAMITE models now support multiple kinematic data sets of different types and with different histogram settings
 - Improvement: the orbit_plot now allows custom orbit cuts via the ocut parameter (same behavior as in the Decomposition class)
 - Bugfix: fixed a crash of the SpecificModels parameter generator with certain specific value lists.